backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# For AI integration with existing hack2 services
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0 

# Authentication dependencies
PyJWT>=2.8.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Redis and Caching
redis>=5.0.0
redis-py-cluster>=2.1.3
aioredis>=2.0.1
fastapi-cache2[redis]>=0.2.1

# Background Tasks
celery>=5.3.0
celery[redis]>=5.3.0
flower>=2.0.0

# Performance and Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0
cachetools>=5.3.0 